cache-api: Avoid offset modification of nl_cache_ops
Commit 18309aedf52 has introduces a new member co_refcnt
(cache: Add reference counter to cache operations) which is
strictly required in order to provide thread safe cache ops
handling.

We have no choice but to modify nl_cache_ops but we might as
well do it at the end of the struct in order to minimize
ABI breakage to a minimum.

Also, adding some space for future expansion if we are beaking
ABI anyway.

Signed-off-by: Thomas Graf <tgraf@suug.ch>